test(system): add basic test setup
