+++ .eslintrc
@@ -1,3 +1,16 @@
 {
-  "extends": ["xo/esnext"]
+  "root": true,
+  "extends": ["xo/esnext"],
+  "rules": {
+    "max-len": [
+      "error",
+      {
+        "code": 80,
+        "comments": 100,
+        "ignoreTrailingComments": true,
+        "ignoreUrls": true,
+        "tabWidth": 2
+      }
+    ]
+  }
 }
