+++ package.json
@@ -12,8 +12,10 @@
     "prepare": "mkdir -p distribution",
     "prebuild": "parallelshell 'npm run test' 'npm run clean && npm run prepare'",
     "build": "babel source --out-dir distribution",
-    "test": "eslint source/**/*.js && jsonlint-cli **/*.json",
-    "watch": "npm run build -- --watch",
+    "test": "parallelshell 'npm run test:ava' 'npm run test:lint'",
+    "test:ava": "ava",
+    "test:lint": "parallelshell 'eslint source/**/*.js' '&& jsonlint-cli **/*.json'",
+    "watch": "npm run test:ava -- --watch",
     "commit": "git-cz",
     "commitmsg": "conventional-changelog-lint -e",
     "changelog": "conventional-changelog --preset angular --infile changelog.md --same-file --output-unreleased",
@@ -27,6 +29,18 @@
     "commitizen": {
       "path": "cz-conventional-changelog-lint"
     }
+  },
+  "ava": {
+    "files": [
+      "test/**/*.js"
+    ],
+    "source": [
+      "source/**/*.js"
+    ],
+    "require": [
+      "babel-register"
+    ],
+    "babel": "inherit"
   },
   "repository": {
     "type": "git",
@@ -58,25 +72,32 @@
   ],
   "license": "MIT",
   "devDependencies": {
+    "ava": "0.13.0",
     "babel": "6.5.2",
     "babel-cli": "6.4.0",
     "babel-core": "5.8.35",
-    "babel-eslint": "4.1.8",
+    "babel-eslint": "6.0.0",
     "babel-plugin-add-module-exports": "0.1.2",
+    "babel-plugin-transform-async-to-module-method": "6.7.0",
+    "babel-preset-async-to-bluebird": "1.0.0",
     "babel-preset-es2015": "6.3.13",
+    "babel-preset-es2015-node5": "1.2.0",
+    "babel-preset-es2016-node5": "1.1.2",
     "babel-preset-stage-0": "6.3.13",
+    "babel-register": "6.7.2",
     "commitizen": "2.7.3",
     "conventional-changelog-cli": "1.1.1",
     "conventional-changelog-lint": "0.3.2",
     "conventional-recommended-bump": "0.1.0",
     "cz-conventional-changelog": "1.1.5",
     "cz-conventional-changelog-lint": "0.1.2",
-    "eslint": "1.10.3",
-    "eslint-config-xo": "0.9.1",
+    "eslint": "2.5.3",
+    "eslint-config-xo": "0.12.0",
     "eslint-plugin-babel": "3.1.0",
     "husky": "0.10.2",
     "jsonlint-cli": "1.0.1",
-    "parallelshell": "2.0.0"
+    "parallelshell": "2.0.0",
+    "unindent": "2.0.0"
   },
   "peerDependencies": {
     "babel-core": "^5",
