+++ source/function-template.js
@@ -1,0 +1,10 @@
+import template from 'babel-template';
+
+export default template(`
+	function NAME (props) {
+		AUXILIARY;
+		return (JSX);
+	}
+`, {
+	sourceType: 'module'
+});
