+++ source/import-template.js
@@ -1,0 +1,5 @@
+import template from 'babel-template';
+
+export default template(`const LOCAL = require(IMPORTED);`, {
+	sourceType: 'module'
+});
