+++ source/normalize-tag-name.js
@@ -1,0 +1,5 @@
+export default function normalizeTagName(tagName) {
+	return tagName.toUpperCase() === tagName ?
+		tagName.toLowerCase() :
+		tagName;
+}
