+++ source/outermost-scope-types.js
@@ -1,0 +1,1 @@
+export default ['Program', 'File'];
