+++ source/supports-stateless-components.js
@@ -1,0 +1,5 @@
+import compareSemver from 'compare-semver';
+
+export default function supportsStatelessComponents(React) {
+	return compareSemver.gt(React.version, ['0.13.3']);
+}
